backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
databases==0.9.0
deptry==0.20.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
lightgbm==4.5.0
numpy==2.1.2
pandas==2.2.3
peewee==3.17.6
pip-autoremove==0.10.0
pkce==1.0.3
pkcegen==1.0.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
simple-jwt-decode==0.9.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
